Simplify js files
