--- assets/src/js/acf-experiments.ts
@@ -1,10 +1,0 @@
-/**
- * Experiments TypeScript
- *
- * @package    Secure Custom Fields
- * @subpackage Admin
- * @since      6.4.3
- */
-
-// Import experiments module
-import './_acf-experiments';
